Unicode::Normalize:illegal.t: Use smaller code point
This commit changes the test to use a code point that isn't
deprecated on any platform.  Some large code points are being
deprecated, and for the purposes of this test, we just need a large code
point, the exact value of which is immaterial, so we can choose one that
won't cause problems.