Fix mkheader to work on releases earlier than 5.8
utf8::upgrade() didn't come along until 5.8, so this failed on earlier
releases.  I've split the pack_U() function into 2 versions, where
the one for earlier releases uses pack("U*"), and the later uses
pack(W*), appended by a pack(U) to force the upgrade.  (Of cource an
explicit utf8::upgrade() could have been used for the later version, but
this was in keeping with the earlier design)

The empty pack("U") now appears to me to have been added for this reason
in commits dating from that time.  I was apparently wrong to think it
had anything to do with EBCDIC, so have changed the comment about that
in Normalize.pm